+++ src/main/java/com/twilio/rest/Domains.java
@@ -10,8 +10,8 @@ package com.twilio.rest;
 public enum Domains {
     ACCOUNTS("accounts"),
     API("api"),
-    CHAT("ip-messaging"),
-    IPMESSAGING("ip-messaging"),
+    CHAT("chat"),
+    IPMESSAGING("chat"),
     LOOKUPS("lookups"),
     MONITOR("monitor"),
     PRICING("pricing"),